Compute the sample playback step in double
The step a voice advances its cursor by was computed entirely in float:
sampleRate / outputRate scaled by a powf of the pitch correction.  Upstream
computes it as a number.  The stored field had already been widened, so the
double simply held float32's answer -- 0.918749988 where upstream has
0.918750000.

A relative pitch error of about one part in a hundred million sounds like
nothing, and on a short note it is.  On a sustained one the cursor drifts, and
two renders that started in phase end out of it, which is why this showed up as
a synthesis difference rather than a tuning one.

It is the largest single fix on this branch: 26 of 31 corpus files improve,
none regress, mean +21.97 dB.  stereo_eq 52.92 -> 137.09, auto_pan
65.12 -> 141.60, phaser_manual_param 71.47 -> 135.91, tremolo_efx_test
74.01 -> 134.94, auto_wah 54.55 -> 109.91, gs_chorus_level 60.47 -> 109.83,
gs_filter_test 87.96 -> 130.02, phaser 48.46 -> 78.81.  Near-identical goes
from eighteen files to twenty-six and only two majors remain.

It also explains the three insertion effects that would not respond to widening
their own arithmetic: they were never insertion bugs.  Every grid axis gains
one to two dB with them, and the four C tests pass.

+++ spessasynth_core/src/synthesizer/audio_engine/channel/note_on.c
@@ -346,7 +346,7 @@ void ss_channel_note_on_ex(SS_MIDIChannel *ch, int note, int vel, double time, b
 
 		/* Playback step */
 		uint32_t sr = proc ? proc->sample_rate : 44100;
-		audio.playback_step = (float)samp->sample_rate / (float)sr * powf(2.0f, (float)samp->pitch_correction / 1200.0f);
+		audio.playback_step = (double)samp->sample_rate / (double)sr * pow(2.0, (double)samp->pitch_correction / 1200.0);
 
 		/* Velocity override */
 		int voice_vel = real_vel;
